refactor(log): split log.ts god-file (DEBT #4)
Mechanical, behaviour-identical split of electron/shared/log.ts (637 LOC)
into cohesive siblings; log.ts keeps the public logger facade + re-imports:

  - logRuntime.ts  — isElectronRuntime/getElog/getApp lazy resolvers, types,
                     the single _elog cache singleton
  - logState.ts    — env-flag reads + sqlite-persisted level read/write,
                     LOG_SCHEMA_VERSION
  - logFormat.ts   — jsonlFormat, buildHeader, ensureHeader
  - logRotation.ts — archiveMainLog (10-file) + archiveRendererLog (2-file)

No runtime/format/redaction change. The _currentLevel / _initialized /
_fileSinkEnabled singletons stay in the facade (one source of truth);
buildHeader/ensureHeader take currentLevel as a param.
